Use relaxed TokenRepositoryInterface in BearerTokenAuthorization
The deprecation of BearerTokenRepositoryInterface in favor of
TokenRepositoryInterface requires people implementing the interface
to change their implementation.

However if they do that and swap the interfaces, they can no longer
use BearerTokenAuthorization in their API calls. The library is
basically forcing everyone to keep using deprecated interface.

This commit fixes the issue and allows to use more relaxed
TokenRepositoryInterface in BearerTokenAuthorization. This is backwards
compatible change, because anyone implementing BearerTokenRepositoryInterface
is already implementing TokenRepositoryInterface.